Rewrite join_game presenter and events
