+++ love_letter/models/__init__.py
@@ -8,7 +8,7 @@ from typing import Dict, List, Optional, Union
 from pydantic import BaseModel
 
 from love_letter.models.cards import Card, Deck, PriestCard, find_card_by_name
-from love_letter.models.event import GameCreatedEvent
+from love_letter.models.event import ExceptionEvent, GameCreatedEvent
 from love_letter.models.exceptions import GameException
 
 num_of_player_with_tokens_to_win = {2: 7, 3: 5, 4: 4}
@@ -141,7 +141,7 @@ class Game:
 
     def join(self, player: "Player") -> Optional[GameCreatedEvent]:
         if self.has_started():
-            raise GameException("Game Has Started")
+            raise GameException(ExceptionEvent(message="Game Has Started"))
         # TODO it is no way to verify two players with same name, just pass it
         join_before = [p for p in self.players if p.name == player.name]
         if join_before:
@@ -149,9 +149,11 @@ class Game:
 
         if len(self.players) < 4:
             self.players.append(player)
-            return GameCreatedEvent(game_id=self.id)
+            return GameCreatedEvent(game_id=self.id, success=True)
 
-        raise GameException("Game Has No Capacity For New Players")
+        raise GameException(
+            ExceptionEvent(message="Game Has No Capacity For New Players")
+        )
 
     def start(self):
         if len(self.players) < 2:
