+++ love_letter/models/event.py
@@ -8,3 +8,12 @@ class DomainEvent:
 @dataclass
 class GameCreatedEvent(DomainEvent):
     game_id: str
+    success: bool
+
+
+@dataclass
+class ExceptionEvent(DomainEvent):
+    message: str
+
+    def __str__(self):
+        return self.message
